+++ src/js/Progress.d.ts
@@ -1,3 +1,12 @@
+type DateInterval = {
+    y: number;
+    m: number;
+    d: number;
+    h: number;
+    i: number;
+    s: number;
+    f: number;
+};
 type EventOptions = {
     updateIntervalMSThreshold?: number;
 };
@@ -11,23 +20,11 @@ export default class Progress {
     private _unit;
     constructor(totalCount?: number | null, locale?: string | null, unit?: string | null);
     get Counter(): number;
-    get ElapsedTime(): {
-        y: number;
-        m: number;
-        d: number;
-        h: number;
-        i: number;
-        s: number;
-        f: number;
-    };
+    get ElapsedTime(): DateInterval;
     get PercentageCompleted(): number | null;
     get StartTime(): Date;
     get TotalCount(): number | null;
-    calculateEstimatedTimeEnroute(): {
-        hours: number;
-        minutes: number;
-        seconds: number;
-    } | null;
+    calculateEstimatedTimeEnroute(): DateInterval | null;
     calculateEstimatedTimeOfArrival(): Date | null;
     private formatValue;
     incrementCounter(): void;