+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- tests
+
+### Fixed
+
+- calculation for EstimatedTimeEnroute
+
 ## [1.1.1] - 2024-03-22
 
 ### Fixed